+++ src/main/scala/ValExec_MulAddRecFN.scala
@@ -109,7 +109,7 @@ class ValExec_MulAddRecFN_add(expWidth: Int, sigWidth: Int) extends Module
     mulAddRecFN.io.op := UInt(0)
     mulAddRecFN.io.a := recFNFromFN(expWidth, sigWidth, io.a)
     mulAddRecFN.io.b :=
-        UInt(1, expWidth + sigWidth + 1)<<UInt(expWidth + sigWidth - 1)
+        UInt(1, expWidth + sigWidth + 1)<<(expWidth + sigWidth - 1)
     mulAddRecFN.io.c := recFNFromFN(expWidth, sigWidth, io.b)
     mulAddRecFN.io.roundingMode := io.roundingMode
 
@@ -155,7 +155,7 @@ class ValExec_MulAddRecFN_mul(expWidth: Int, sigWidth: Int) extends Module
     mulAddRecFN.io.b := recFNFromFN(expWidth, sigWidth, io.b)
     mulAddRecFN.io.c :=
         ((io.a ^ io.b) &
-             UInt(1, expWidth + sigWidth + 1)<<UInt(expWidth + sigWidth - 1))<<
+             UInt(1, expWidth + sigWidth + 1)<<(expWidth + sigWidth - 1))<<
             UInt(1)
     mulAddRecFN.io.roundingMode := io.roundingMode
